+++ dalia/__init__.py
@@ -1,1 +1,5 @@
+# flake8: noqa
 __version__ = '0.0.1a1'
+
+
+from dalia.csvv import read_csvv