+++ HISTORY.rst
@@ -1,6 +1,16 @@
+.. currentmodule:: dalia
+
 =======
 History
 =======
+
+
+0.0.1a2 (2019-10-10)
+--------------------
+
+* Add :py:func:`read_csvv` and :py:class:`dalia.csvv.Csvv` to read and represent CSVV files.
+* Initial documentation.
+
 
 0.0.1a1 (2019-10-10)
 --------------------